Added navbar and refactored code
