+++ backend/bingo/templates/bingo/bingo.html
@@ -69,5 +69,5 @@
 <script src="{% static 'js/bingo/BingoSec.js' %}"></script>
 <script src="{% static 'js/bingo/Bingo.js' %}"></script>
 <script src="{% static 'js/bingo/BingoSocket.js' %}"></script>
-<script src="{% static 'js/Utils.js' %}"></script>
+<script src="{% static 'js/utils.js' %}"></script>
 {% endblock script %}
