+++ backend/bingo/templates/bingo/create_room.html
@@ -38,5 +38,5 @@
   href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.1.1/animate.min.css"
 />
 <script src="{% static 'js/bingo/CreateRoom.js' %}"></script>
-<script src="{% static 'js/Utils.js' %}"></script>
+<script src="{% static 'js/utils.js' %}"></script>
 {% endblock script %}
