+++ backend/tictactoe/templates/tictactoe/home.html
@@ -40,5 +40,5 @@
 
 {% endblock content %}
 {% block script %}
-<script src="{% static 'js/Utils.js' %}"></script>
+<script src="{% static 'js/utils.js' %}"></script>
 {% endblock script %}
