+++ backend/tictactoe/templates/tictactoe/play.html
@@ -34,5 +34,5 @@
   </script>
   <script src="https://unpkg.com/sweetalert/dist/sweetalert.min.js"></script>
   <script src="{% static 'js/tictactoe/Play.js' %}"></script>
-  <script src="{% static 'js/Utils.js' %}"></script>
+  <script src="{% static 'js/utils.js' %}"></script>
 {% endblock script %}
