+++ frontend/src/components/atoms/Card/Card.css
@@ -51,7 +51,6 @@ p {
 }
 
 a {
-  text-transform: uppercase;
   color: #1b9cfc;
   text-decoration: none;
   font-weight: bold;
