+++ frontend/src/components/organisms/Footer/Footer.js
@@ -1,0 +1,7 @@
+export default function Footer() {
+  return (
+    <div>
+      <p></p>
+    </div>
+  );
+}
