+++ frontend/src/utils/consts.js
@@ -1,7 +1,8 @@
-const backendHome = 'localhost:8000';
+const backendHome = 'http://localhost:8000';
 
 export const PATHS = {
   home: '/',
+  monitoring: '/monitoring/',
   whiteboard: '/whiteboard/',
   bingo: `${backendHome}/bingo/`,
   tictactoe: `${backendHome}/tictactoe/`
